+++ src/pages/ListagemMarcas/index.js
@@ -1,0 +1,3 @@
+import ListagemMarcas from "./ListagemMarcas";
+
+export default ListagemMarcas;
